6011번: 실수를 입력받아 float형으로 출력

+++ python/codeup/inoutput_6009.py
@@ -8,7 +8,14 @@ if __name__ == '__main__':
 
 
 
-    # 10.숫자 입력받아 int형ㅡ로 출력
+    # 10.숫자 입력받아 int형ㅡ로 출력(6010)
+    # num = input()
+    # num = int(num)
+    # print(num)
+
+
+
+    # 11.실수 입력 받아 출력(6011)
     num = input()
-    num = int(num)
+    num = float(num)
     print(num)
